requirements.txt
pytest
selenium
requests
pytest-html-reporter
allure-pytest